+++ Software/MATLAB/app/system/+matlab/+compiler/+mlspark/+functions/monotonically_increasing_id.m
@@ -1,21 +1,17 @@
-function col = year(inCol)
-    % YEAR Create a numeric year column from a datetime column
+function col = monotonically_increasing_id()
+    % monotonically_increasing_id Generate a column with increasing ID values.
     %
-    % This function will return a new column
+    % This function will return a new column with integer values starting at 0.
     %
     % Example:
     %
-    %     % DS is a dataset
-    %     % Get datetime column
-    %     dtc = DS.col("date")
-    %     % Convert this to a column with year values
-    %     mc = year(dtc)
+    %     % Add a column with numeric ID values to dataset DS
+    %     newDS = DS.withColumn('ID',monotonically_increasing_id())
 
     % Copyright 2020-2022 MathWorks, Inc.
 
     try
-        try inCol = inCol.column; catch, end  % col may be a column name or object
-        jcol = org.apache.spark.sql.functions.year(inCol);
+        jcol = org.apache.spark.sql.functions.monotonically_increasing_id();
     catch err
         error('SPARK:ERROR', 'Spark error: %s', stripJavaError(err.message));
     end