+++ Software/MATLAB/app/system/+matlab/+compiler/+mlspark/+functions/quarter.m
@@ -1,5 +1,5 @@
-function col = year(inCol)
-    % YEAR Create a numeric year column from a datetime column
+function col = quarter(inCol)
+    % QUARTER Create a numeric quarter column from a datetime column
     %
     % This function will return a new column
     %
@@ -8,14 +8,14 @@ function col = year(inCol)
     %     % DS is a dataset
     %     % Get datetime column
     %     dtc = DS.col("date")
-    %     % Convert this to a column with year values
-    %     mc = year(dtc)
+    %     % Convert this to a column with numeric quarter values
+    %     mc = quarter(dtc)
 
     % Copyright 2020-2022 MathWorks, Inc.
 
     try
         try inCol = inCol.column; catch, end  % col may be a column name or object
-        jcol = org.apache.spark.sql.functions.year(inCol);
+        jcol = org.apache.spark.sql.functions.quarter(inCol);
     catch err
         error('SPARK:ERROR', 'Spark error: %s', stripJavaError(err.message));
     end