+++ Software/MATLAB/app/system/+matlab/+compiler/+mlspark/+functions/second.m
@@ -1,19 +1,29 @@
-function col = second(dateCol)
-    % SECOND Create a second column from a datetime column
+function col = second(inCol)
+    % SECOND Create a numeric seconds column from a datetime column
     %
-    % This function will return a new dataset, with the additional column
+    % This function will return a new column
+    %
     % Example:
     %
-    %     % DS is a datasetf
+    %     % DS is a dataset
     %     % Get datetime column
     %     dtc = DS.col("date")
-    %     % Convert this to a date on the dataset
+    %     % Convert this to a column with numeric seconds values
     %     mc = second(dtc)
-    
-    %                 Copyright 2020 MathWorks, Inc.
-    
-    col = matlab.compiler.mlspark.Column(...
-        org.apache.spark.sql.functions.second(dateCol.column) ...
-        );
 
-end %function
+    % Copyright 2020-2022 MathWorks, Inc.
+
+    try
+        try inCol = inCol.column; catch, end  % col may be a column name or object
+        jcol = org.apache.spark.sql.functions.second(inCol);
+    catch err
+        error('SPARK:ERROR', 'Spark error: %s', stripJavaError(err.message));
+    end
+    if ~isempty(jcol)
+        col = matlab.compiler.mlspark.Column(jcol);
+    else
+        error('SPARK:ERROR', ...
+            'The Spark %s function only supports an argument that is a matlab.compiler.mlspark.Column object or a column name', ...
+            mfilename);
+    end
+end